Protect feed provider config from cache clearing
